+++ CHANGELOG.md
@@ -5,6 +5,17 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+## [0.2.0] - 2019-06-17
+
+### Added
+- `dblocalization.enabled` configuration property to enable DB based message lookup
+- support for Postgres DBMS
+
+### Bugfix
+- correct lookup for frontend message keys
+
+### Dependencies
+- CUBA 7.0.x
 ## [0.1.1] - 2019-06-16
 
 ### Added